Empty commit; note
For simplicity's sake, let's assume that there is at least a single
pixel column that's completely clear from each character to the next,
so we don't need to look at the previous or next character, only the
previous and next column of the current character.

Even if that turns out not to be the case, I'm happy to let this
result in (for each whole character) the clearly demarked columning
you currently get with each and every column.